image_parser/requirements.txt
kafka-python==2.0.2
numpy==1.21.1
Pillow==9.0.1
opencv-python==4.7.0.68
tensorflow==2.11.0
minio==7.1.12
pydantic==1.9.0
requests==2.27.1